*** CAN: Introduce listen-only mode
Listen-only mode allows receiving CAN frames from a bus without
disturbing the bus.

This parameter can be set by a configuration key.

PCAN_LISTEN_ONLY:
This parameter is a so called "pre-initialized" parameter, which means
that it can be set before a PCAN-Channel is initialized in order to
activate/deactivate the parameter as fast as possible, avoiding in this
way problems that can appears within sensitive operations.

XL_OUTPUT_MODE:
To call this function, the port must have init access (see section
xlOpenPort on page 37) for  the specified channels, and the  channels
must be deactivated.

Change-Id: I81277cfbffe96a8793a45c0c30e291a5577246cf